use a user-agent for requests (but not youtube)
Add some more hacks because why not. By the way, mechanize is used
because requests sucks at logging into forums and randomly breaks.
However, mechanize sucks with certain domain names for no reason and
randomly breaks so requests is used instead. Here, we set a user agent
to look like less of a robot for some sites but youtube seems to screw
up more (but it still screws up sometimes) if we use a user agent. Who
knows. Webshit is terrible.