execute_command: refactor to more sanity
The old ways horrible and dumb. This way is less horrible and dumb. Use
a big overaching execute_command function with little sub functions in
each if branch. In there, we actually split the string into the
arguments needed for the command. Additionally, clean up how the command
string is split up throughout the entire process to be more clear and
less arbitrary. Rely on each function if branch to split correctly
instead of the weird convoluted way things were being split before. This
makes it simplier to add new commands.